requirements.txt
aiofiles==23.2.1 ; python_version >= "3.11" and python_version < "3.12"
aiohttp==3.9.5 ; python_version >= "3.11" and python_version < "3.12"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "3.12"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "3.12"
anyio==4.4.0 ; python_version >= "3.11" and python_version < "3.12"
async-lru==2.0.4 ; python_version >= "3.11" and python_version < "3.12"
async-timeout==4.0.3 ; python_version >= "3.11" and python_full_version <= "3.11.2"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "3.12"
av==10.0.0 ; python_version >= "3.11" and python_version < "3.12"
boto3==1.34.113 ; python_version >= "3.11" and python_version < "3.12"
botocore==1.34.113 ; python_version >= "3.11" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "3.12" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.12" and (sys_platform == "win32" or platform_system == "Windows")
coloredlogs==15.0.1 ; python_version >= "3.11" and python_version < "3.12"
cryptography==42.0.7 ; python_version >= "3.11" and python_version < "3.12"
ctranslate2==3.24.0 ; python_version >= "3.11" and python_version < "3.12"
dataclasses-json==0.6.6 ; python_version >= "3.11" and python_version < "3.12"
distro==1.9.0 ; python_version >= "3.11" and python_version < "3.12"
fastapi-versionizer==3.0.4 ; python_version >= "3.11" and python_version < "3.12"
fastapi==0.109.0 ; python_version >= "3.11" and python_version < "3.12"
faster-whisper==0.10.1 ; python_version >= "3.11" and python_version < "3.12"
filelock==3.14.0 ; python_version >= "3.11" and python_version < "3.12"
flatbuffers==24.3.25 ; python_version >= "3.11" and python_version < "3.12"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "3.12"
fsspec==2024.5.0 ; python_version >= "3.11" and python_version < "3.12"
greenlet==3.0.3 ; python_version >= "3.11" and python_version < "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.11" and python_version < "3.12"
httpcore==1.0.5 ; python_version >= "3.11" and python_version < "3.12"
httptools==0.6.1 ; python_version >= "3.11" and python_version < "3.12"
httpx==0.27.0 ; python_version >= "3.11" and python_version < "3.12"
huggingface-hub==0.23.2 ; python_version >= "3.11" and python_version < "3.12"
humanfriendly==10.0 ; python_version >= "3.11" and python_version < "3.12"
idna==3.7 ; python_version >= "3.11" and python_version < "3.12"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "3.12"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
jsonpatch==1.33 ; python_version >= "3.11" and python_version < "3.12"
jsonpointer==2.4 ; python_version >= "3.11" and python_version < "3.12"
langchain-community==0.0.20 ; python_version >= "3.11" and python_version < "3.12"
langchain-core==0.1.23 ; python_version >= "3.11" and python_version < "3.12"
langchain-openai==0.0.6 ; python_version >= "3.11" and python_version < "3.12"
langchain==0.1.7 ; python_version >= "3.11" and python_version < "3.12"
langsmith==0.0.87 ; python_version >= "3.11" and python_version < "3.12"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "3.12"
marshmallow==3.21.2 ; python_version >= "3.11" and python_version < "3.12"
mpmath==1.3.0 ; python_version >= "3.11" and python_version < "3.12"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "3.12"
natsort==8.4.0 ; python_version >= "3.11" and python_version < "3.12"
networkx==3.3 ; python_version >= "3.11" and python_version < "3.12"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.12"
onnxruntime==1.18.0 ; python_version >= "3.11" and python_version < "3.12"
openai==1.30.3 ; python_version >= "3.11" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.11" and python_version < "3.12"
prometheus-client==0.19.0 ; python_version >= "3.11" and python_version < "3.12"
prometheus-fastapi-instrumentator==6.1.0 ; python_version >= "3.11" and python_version < "3.12"
protobuf==5.27.0 ; python_version >= "3.11" and python_version < "3.12"
pycparser==2.22 ; python_version >= "3.11" and python_version < "3.12" and platform_python_implementation != "PyPy"
pydantic-core==2.14.6 ; python_version >= "3.11" and python_version < "3.12"
pydantic==2.5.3 ; python_version >= "3.11" and python_version < "3.12"
pyjwt[crypto]==2.8.0 ; python_version >= "3.11" and python_version < "3.12"
pyreadline3==3.4.1 ; sys_platform == "win32" and python_version >= "3.11" and python_version < "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "3.12"
redis==5.0.1 ; python_version >= "3.11" and python_version < "3.12"
regex==2024.5.15 ; python_version >= "3.11" and python_version < "3.12"
requests==2.32.2 ; python_version >= "3.11" and python_version < "3.12"
s3transfer==0.10.1 ; python_version >= "3.11" and python_version < "3.12"
setuptools==70.0.0 ; python_version >= "3.11" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.11" and python_version < "3.12"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy==2.0.30 ; python_version >= "3.11" and python_version < "3.12"
starlette==0.35.1 ; python_version >= "3.11" and python_version < "3.12"
sympy==1.12 ; python_version >= "3.11" and python_version < "3.12"
tenacity==8.3.0 ; python_version >= "3.11" and python_version < "3.12"
tiktoken==0.7.0 ; python_version >= "3.11" and python_version < "3.12"
tokenizers==0.15.2 ; python_version >= "3.11" and python_version < "3.12"
torch==2.0.1 ; python_version >= "3.11" and python_version < "3.12"
torchaudio==2.0.2 ; python_version >= "3.11" and python_version < "3.12"
tqdm==4.66.4 ; python_version >= "3.11" and python_version < "3.12"
typing-extensions==4.12.0 ; python_version >= "3.11" and python_version < "3.12"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "3.12"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "3.12"
uuid6==2024.1.12 ; python_version >= "3.11" and python_version < "3.12"
uvicorn[standard]==0.29.0 ; python_version >= "3.11" and python_version < "3.12"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "3.12"
watchfiles==0.22.0 ; python_version >= "3.11" and python_version < "3.12"
websockets==12.0 ; python_version >= "3.11" and python_version < "3.12"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "3.12"
